requirements.txt
Pillow==8.2.0
tesserocr==2.5.1
